Modified interface, added tests, refactored trainer interface due to a previous method name change - this may cause conflicts. Relocated the web  driver tests to a more appropriate package.
