+++ src/test/java/com/sparta/eng82/interfaces/pages/navpages/trainee/TraineeHomePage.java
@@ -27,4 +27,6 @@ public interface TraineeHomePage extends NavPage {
     String getGroup();
 
     String getTrainer();
+
+    String getMaxWeek();
 }
