+++ src/test/java/com/sparta/eng82/tests/unit/utility/WebDriverTests.java
@@ -1,5 +1,7 @@
-package com.sparta.eng82.components.webdriver;
+package com.sparta.eng82.tests.unit.utility;
 
+import com.sparta.eng82.components.webdriver.WebDriverFactory;
+import com.sparta.eng82.components.webdriver.WebDriverTypes;
 import org.junit.jupiter.api.*;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
